+++ packages/superset-ui-chart/src/index.ts
@@ -16,14 +16,6 @@ export {
   default as getChartTransformPropsRegistry,
 } from './registries/ChartTransformPropsRegistrySingleton';
 
-export { default as buildQueryContext } from './query/buildQueryContext';
-export { default as DatasourceKey } from './query/DatasourceKey';
-
 export { default as ChartDataProvider } from './components/ChartDataProvider';
 
-export * from './types/Annotation';
-export * from './types/ChartFormData';
-export * from './types/Datasource';
-export * from './types/Metric';
-export * from './types/Query';
 export * from './types/TransformFunction';
