+++ packages/superset-ui-query/src/buildQueryObject.ts
@@ -1,6 +1,6 @@
 /* eslint-disable camelcase */
-import { QueryObject } from '../types/Query';
-import { ChartFormData, isSqlaFormData } from '../types/ChartFormData';
+import { QueryObject } from './types/Query';
+import { ChartFormData, isSqlaFormData } from './types/ChartFormData';
 import convertMetric from './convertMetric';
 import processFilters from './processFilters';
 import processMetrics from './processMetrics';
@@ -12,11 +12,13 @@ function processGranularity(formData: ChartFormData): string {
   return isSqlaFormData(formData) ? formData.granularity_sqla : formData.granularity;
 }
 
-// Build the common segments of all query objects (e.g. the granularity field derived from
-// either sql alchemy or druid). The segments specific to each viz type is constructed in the
-// buildQuery method for each viz type (see `wordcloud/buildQuery.ts` for an example).
-// Note the type of the formData argument passed in here is the type of the formData for a
-// specific viz, which is a subtype of the generic formData shared among all viz types.
+/**
+ * Build the common segments of all query objects (e.g. the granularity field derived from
+ * either sql alchemy or druid). The segments specific to each viz type is constructed in the
+ * buildQuery method for each viz type (see `wordcloud/buildQuery.ts` for an example).
+ * Note the type of the formData argument passed in here is the type of the formData for a
+ * specific viz, which is a subtype of the generic formData shared among all viz types.
+ */
 export default function buildQueryObject<T extends ChartFormData>(formData: T): QueryObject {
   const {
     time_range,
