+++ packages/superset-ui-query/src/processFilters.ts
@@ -1,6 +1,6 @@
-import { ChartFormData } from '../types/ChartFormData';
-import { QueryObjectFilterClause } from '../types/Query';
-import { isSimpleAdhocFilter } from '../types/Filter';
+import { ChartFormData } from './types/ChartFormData';
+import { QueryObjectFilterClause } from './types/Query';
+import { isSimpleAdhocFilter } from './types/Filter';
 import convertFilter from './convertFilter';
 
 /** Logic formerly in viz.py's process_query_filters */
