+++ packages/superset-ui-query/src/types/Column.ts
